Rename suggested branch name from yourname to avoid confusion
The welcome-demo README file gives an example of how to create and use a
branch.  The branch name in the example has been <yourname>, and some
students apparently have thought this example meant that every branch in
every repo should be named simply with their username.  On the contrary,
branches should have a short but descriptive name (as noted in our
GitHub workflow at
https://projects.csail.mit.edu/cgi-bin/wiki/view/Infolab/GithubWorkflow,
which is referenced in the README but apparently not always read or
remembered).

Change the example branch name to add-<yourname>-fav-num in hopes this
will alleviate the problem.