Say how to enter multiline log message
It's not necessary when going through the Welcome Demo, but we walk
people through the Welcome Demo and then throw them into the deep end
without ever noting that there's a good way to write longer messages.
Note it here in hopes that people will remember when they start making
"real" commits.